Reordered the main files modified manifest.json added base file to future tasks
