+++ manifest.json
@@ -4,7 +4,7 @@
     "description": "Boost wolt up 🚀",
     "manifest_version": 3,
     "permissions": [
-      "storage"
+      "storage", "tabs"
     ],
     "options_ui": {
       "page": "/src/set-team-page.html",
@@ -62,5 +62,14 @@
       "24": "/assets/icons/pimp-my-wolt-icon-24.png",
       "48": "/assets/icons/pimp-my-wolt-icon-48.png",
       "128": "/assets/icons/pimp-my-wolt-icon-128.png"
+    },
+    "action": {
+      "default_icon": {
+        "16": "/assets/icons/pimp-my-wolt-icon-16.png",
+        "24": "/assets/icons/pimp-my-wolt-icon-24.png",
+        "48": "/assets/icons/pimp-my-wolt-icon-48.png"
+      },
+      "default_title": "Pimp My Wolt",
+      "default_popup": "popup.html"
     }
   }
