+++ src/category-random.html
@@ -1,0 +1,11 @@
+<!DOCTYPE html>
+<html>
+<head><title>Pimp my Wolt</title></head>
+<body>
+  
+  <span>Generate a random food category</span> <br/>
+  <input type="button" onclick="rndCategory()"></input>
+  <p id="generated-text"></p>
+  <script src="category-random.js"></script>
+  </body>
+</html>
