+++ src/pimp-my-wolt.js
@@ -1,3 +1,4 @@
+const pmwBL = require('./pmwBL.js');
 (function () {
   const logoUrl = chrome.runtime.getURL(
     "assets/icons/pimp-my-wolt-icon-48.png"
@@ -74,9 +75,9 @@
     const text = isHebrewWolt ? `הזמן את ${teamName}` : `Invite ${teamName}`;
 
     const onclick = async () => {
-      const { invitedGuests, notInvitedGuests } = await inviteAllGuests();
+      const { invitedGuests, notInvitedGuests } = await pmwBL.inviteAllGuests();
       biLogger.logEvent("invite_all_group", {
-        restaurant: getRestuarant(),
+        restaurant: pmwBL.getRestuarant(),
         invitedGuests,
         notInvitedGuests,
       });
@@ -106,91 +107,6 @@
     return btn;
   }
 
-  async function inviteAllGuests() {
-    const guests = await allGuests;
-    const invitedGuests = [];
-    const notInvitedGuests = [];
-    for (guest of guests) {
-      const guestName = guest.woltName;
-      const inviteButton = getElementWithText("li", guestName)?.querySelector(
-        "button"
-      );
-      inviteButton?.click();
-      (inviteButton ? invitedGuests : notInvitedGuests).push(guestName);
-    }
-    return {
-      invitedGuests,
-      notInvitedGuests,
-    };
-  }
-
-  function getRestuarant() {
-    const location = document?.location?.pathname;
-    const regex = /restaurant\/(.*)\//gm;
-    const restaurant = location?.matchAll(regex)?.next()?.value?.[1];
-    return restaurant;
-  }
-
-  function getTotalOrderPrice() {
-    const amountWithCurrency = getElementWithText(
-      "dl",
-      totalText
-    )?.querySelector("dd")?.innerText;
-    return priceToNumber(amountWithCurrency ?? "");
-  }
-
-  function priceToNumber(price) {
-    if (typeof price !== "string") return 0;
-    const maybeNumber = Number(price.replace(/[^0-9.-]+/g, ""));
-    return isNaN(maybeNumber) ? 0 : maybeNumber;
-  }
-
-  function getGuestsOrders() {
-    const orderTable = document.querySelector(
-      "[class^=Tabs-module__root] [class^=Tabs-module__content]"
-    );
-    const guestsLineItems = Array.from(
-      orderTable?.querySelectorAll("li") || []
-    );
-
-    return guestsLineItems
-      .map((item) => {
-        const name = item.querySelector(
-          '[class*="GuestItem-module__listName"] span'
-        )?.innerText;
-        const price = priceToNumber(
-          item.querySelector('[class*="GuestItem-module__price"]')?.innerText
-        );
-
-        return {
-          name,
-          price,
-        };
-      })
-      .filter((guest) => guest.name && guest.price);
-  }
-
-  function updateOrderButtonToSaveGuestsOrders() {
-    const sendOrderButton = document.querySelector(
-      '[data-test-id="SendOrderButton"]'
-    );
-    sendOrderButton.onclick = () => {
-      const totalOrderPrice = getTotalOrderPrice();
-      const guestsOrders = getGuestsOrders();
-      const restaurant = getRestuarant();
-      const orderTimestamp = Date.now();
-      chrome.storage.local.set({
-        totalOrderPrice,
-        guestsOrders,
-        orderTimestamp,
-        restaurant,
-      });
-    };
-    sendOrderButton.setAttribute(
-      orderButtonHookSettings.saveOrdersAttribute,
-      "true"
-    );
-  }
 
   async function addInviteGroupButton() {
     const btn = await getBtn();
@@ -413,6 +329,7 @@
     addSetGroupModal();
     addMemberSetupModal();
     addMemberRemovalModal();
+    generateCategoryModal(); //TODO - Write this function 
   }
 
   setInterval(async () => {
@@ -423,7 +340,7 @@
     }
 
     if (isOrderButtonExists() && !isOrderButtonUpdated()) {
-      updateOrderButtonToSaveGuestsOrders();
+      pmwBL.updateOrderButtonToSaveGuestsOrders();
     }
 
     if (
